[librpbase] RomMetaData.hpp: Reduce the `Property` enum from `int` to `int8_t`.
There's currently only 79 properties defined, so `int` wastes space.

This reduces the size of RomMetaDataPrivate from 344 bytes to 104 bytes
(on amd64).

TODO: Don't use an std::array<> for map_metaData?

+++ src/librpbase/RomMetaData.hpp
@@ -25,7 +25,7 @@ namespace LibRpBase {
 
 // Properties.
 // This matches KFileMetaData::Property.
-enum class Property : int {
+enum class Property : int8_t {
 	Invalid = -1,
 	FirstProperty = 0,
 	Empty = 0,
